Adding more stuff

+++ src/main/java/controller/WSStudent.java
@@ -2,6 +2,7 @@ package controller;
 
 //endpoint: http://localhost:8080/ws-sii-patm/---/[controller]
 
+import model.Access;
 import model.Student;
 
 import javax.ws.rs.*;
@@ -12,32 +13,47 @@ import java.util.List;
 public class WSStudent {
 
     @POST
+    @Path("/{token}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public Student insert(Student student) {
-        student.insert();
+    public Student insert(Student student, @PathParam("token") String token) {
+        Access access = new Access();
+        if(access.validateToken(token))
+            student.insert();
+        else
+            student.setName("Acceso denegado");
         return student;
     }
 
     @PUT
+    @Path("/{token}")
     @Consumes(MediaType.APPLICATION_JSON)
-    public void update(Student student) {
-        student.update();
+    public void update(Student student, @PathParam("token") String token) {
+        Access access = new Access();
+        if(access.validateToken(token))
+            student.update();
     }
 
     @DELETE
-    @Path("/{student_id}")
-    public void delete(@PathParam("student_id") int studentId) {
+    @Path("/{student_id}/{token}")
+    public void delete(@PathParam("student_id") int studentId, @PathParam("token") String token) {
         Student student = new Student();
-        student.setStudentId(studentId);
-        student.delete();
+        Access access = new Access();
+        if(access.validateToken(token)) {
+            student.setStudentId(studentId);
+            student.delete();
+        }
     }
 
     @GET
-    @Path("/lista")
+    @Path("/lista/{token}")
     @Produces(MediaType.APPLICATION_JSON)
-    public List<Student> getListOfStudents() {
+    public List<Student> getListOfStudents(@PathParam("token") String token) {
         Student student = new Student();
-        return student.getStudents();
+        Access access = new Access();
+        if(access.validateToken(token))
+            return student.getStudents();
+        else
+            return null;
     }
 }
